building with m68k-elf-gcc (move to C11)

+++ kernel.h
@@ -1,8 +1,9 @@
+#include <stdint.h>
+#include <stddef.h>
+#include "kernel_constants.h"
+
 #ifndef KERNEL_H
 #define KERNEL_H
-
-#include <stdint.h>
-#include "kernel_constants.h"
 
 void kinit();
 void kmain();
@@ -15,7 +16,7 @@ uint8_t  peekb(uint32_t address);
 uint16_t peekw(uint32_t address);
 uint32_t peekl(uint32_t address);
 
-uint8_t *memcpy(uint8_t *dest, const uint8_t *src, uint32_t count);
-uint8_t *memset(uint8_t *dest, uint8_t val, uint32_t count);
+uint8_t *memcpy(uint8_t *dest, const uint8_t *src, size_t count);
+uint8_t *memset(uint8_t *dest, uint8_t val, size_t count);
 
 #endif
